mini-pytest-api-tests/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-pytest-api-tests"
version = "0.1.0"
description = "A production-quality pytest API testing framework example"
authors = [
  { name = " QA Engineer", email = "qa@example.com" },
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "pytest-html>=4.1.1",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.15",
    "black>=24.1.1",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

[tool.pytest.ini_options]
addopts = "-v --html=.artifacts/report.html --self-contained-html"
testpaths = ["tests"]
markers = [
    "smoke: Critical path tests",
    "regression: Full regression suite",
    "slow: Tests that take longer to run",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B"]
ignore = []

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
